pyproject.toml
[tool.poetry]
name = "weissuehotoffers"
version = "0.1.0"
description = ""
authors = ["JackQweata <dahil3@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
selenium = "^4.9.0"
webdriver-manager = "^3.8.6"
fake-headers = "^1.0.2"
python-decouple = "^3.8"
pytelegrambotapi = "^4.11.0"
sqlalchemy = "^2.0.12"
pymysql = "^1.0.3"
pillow = "^9.5.0"
svglib = "^1.5.1"
pyppeteer = "^1.0.2"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
